refactor: removed useBundler & config class, naming conventions and more
- useBundler is not needed anymore, if we want tu run rubocop with bundler we should put it on the command line.
- The Config class is a good overprogramming example. It is removed in favour of CommandBuilder, more simple and cool.
- Drunken camel case for helpers.
- Utility functions extracted to helpers.
- SRP aproximation for Rubocop class, we explicitly pass arguments like current directory, etc...
