+++ src/helpers/hasValidScope.js
