+++ src/rubocop/formatters/ErrorFormatter.js
