+++ src/rubocop/helpers/getRuleDocumentation.js
